unsigned to int

+++ src/Container1d.cpp
@@ -23,41 +23,40 @@ pmat::Container1d &pmat::Container1d::operator=(pmat::Container1d &&container) n
    return *this;
 }
 
-pmat::Container1d::Container1d(const unsigned &size, const double &value)
+pmat::Container1d::Container1d(const int &size, const double &value)
     : _ptrVector{std::make_unique<blitz::Array<double, 1>>(size)} {
    *_ptrVector = value;
 }
 
-void pmat::Container1d::resize(const unsigned &size) {
+void pmat::Container1d::resize(const int &size) {
 
-   unsigned oldSize = this->size();
-   // Caution: narrowing from unsigned to int
-   _ptrVector->resizeAndPreserve((int)size);
+   int oldSize = this->size();
+   // Caution: narrowing from int to int
+   _ptrVector->resizeAndPreserve(size);
 
    if (size > oldSize)
-      for (unsigned i{oldSize}; i < size; i++)
-         (*_ptrVector)((int)i) = utils::ZERO;
+      for (int i{oldSize}; i < size; i++)
+         (*_ptrVector)(i) = utils::ZERO;
 }
 
 void pmat::Container1d::pushBack(const double &value) {
-   // Caution: narrowing from unsigned to int
+   // Caution: narrowing from int to int
    auto size{_ptrVector->size()};
-   _ptrVector->resizeAndPreserve((int)size + 1);
-   (*_ptrVector)((int)size) = value;
+   _ptrVector->resizeAndPreserve(size + 1);
+   (*_ptrVector)(size) = value;
 }
 
-void pmat::Container1d::pushBack(const unsigned &initialPosition, const unsigned &step,
-                                 const double &value) {
-   unsigned oldSize = this->size();
-   unsigned inc = std::floor(oldSize / step);
-   unsigned newSize = oldSize + inc;
-   unsigned i = oldSize - 1;
-   unsigned j = newSize - 1;
-   unsigned k = initialPosition + inc;
-   unsigned newStep = step + inc - 1;
+void pmat::Container1d::pushBack(const int &initialPosition, const int &step, const double &value) {
+   int oldSize = this->size();
+   int inc = std::floor(oldSize / step);
+   int newSize = oldSize + inc;
+   int i = oldSize - 1;
+   int j = newSize - 1;
+   int k = initialPosition + inc;
+   int newStep = step + inc - 1;
    this->resize(newSize);
 
-   unsigned n{0};
+   int n{0};
    while (n < inc) {
       if (k == j) {
          this->set(j--, value);
@@ -68,29 +67,28 @@ void pmat::Container1d::pushBack(const unsigned &initialPosition, const unsigned
    }
 }
 
-void pmat::Container1d::exchange(const unsigned &indexA, const unsigned &indexB) {
-   // Caution: narrowing from unsigned to int
-   (*_ptrVector)((int)indexB) =
-       std::exchange((*_ptrVector)((int)indexA), (*_ptrVector)((int)indexB));
+void pmat::Container1d::exchange(const int &indexA, const int &indexB) {
+   // Caution: narrowing from int to int
+   (*_ptrVector)(indexB) = std::exchange((*_ptrVector)(indexA), (*_ptrVector)(indexB));
 }
 
 void pmat::Container1d::ascendingSort() {
    double *data = _ptrVector->data();
-   unsigned size = _ptrVector->size();
+   int size = (int)_ptrVector->size();
    std::sort(data, data + size);
-   (*_ptrVector) = blitz::Array<double, 1>(data, (int)size);
+   (*_ptrVector) = blitz::Array<double, 1>(data, size);
 }
 
 void pmat::Container1d::descendingSort() {
    double *data = _ptrVector->data();
-   unsigned size = _ptrVector->size();
+   int size = (int)_ptrVector->size();
    std::sort(data, data + size, [](double &left, double &right) -> bool { return left > right; });
-   (*_ptrVector) = blitz::Array<double, 1>(data, (int)size);
+   (*_ptrVector) = blitz::Array<double, 1>(data, size);
 }
 
 void pmat::Container1d::insertValues(const int &initialPosition, const double &value,
                                      const int &nRepetitions) {
-   int oldSize = (int)this->size();
+   int oldSize = this->size();
    int newSize = oldSize + nRepetitions;
    int i = oldSize - 1;
    int j = newSize - 1;
